requirements.txt
azure-cli
tqdm